refactor taxonomy entries

+++ android-client/app/src/main/java/de/davidtiede/slrtoolkit/fragments/BibtexEntriesDetailViewPagerFragment.java
@@ -22,9 +22,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 import de.davidtiede.slrtoolkit.ClassificationActivity;
+import de.davidtiede.slrtoolkit.ProjectActivity;
 import de.davidtiede.slrtoolkit.R;
+import de.davidtiede.slrtoolkit.TaxonomiesActivity;
 import de.davidtiede.slrtoolkit.database.Entry;
 import de.davidtiede.slrtoolkit.viewmodels.ProjectViewModel;
+import de.davidtiede.slrtoolkit.viewmodels.TaxonomiesViewModel;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -32,6 +35,7 @@ import de.davidtiede.slrtoolkit.viewmodels.ProjectViewModel;
 public class BibtexEntriesDetailViewPagerFragment extends Fragment {
     ViewPager2 viewPager;
     ProjectViewModel projectViewModel;
+    TaxonomiesViewModel taxonomiesViewModel;
     private TextView noEntriesDetails;
     private FragmentStateAdapter pagerAdapter;
     int repoId;
@@ -51,7 +55,12 @@ public class BibtexEntriesDetailViewPagerFragment extends Fragment {
 
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         projectViewModel = new ViewModelProvider(requireActivity()).get(ProjectViewModel.class);
-        repoId = projectViewModel.getCurrentRepoId();
+        taxonomiesViewModel = new ViewModelProvider(requireActivity()).get(TaxonomiesViewModel.class);
+        if(requireActivity() instanceof ProjectActivity) {
+            repoId = projectViewModel.getCurrentRepoId();
+        } else if(requireActivity() instanceof TaxonomiesActivity) {
+            repoId = taxonomiesViewModel.getCurrentRepoId();
+        }
         viewPager = view.findViewById(R.id.entries_detail_viewpager);
         noEntriesDetails = view.findViewById(R.id.textview_no_entries_details);
         pagerAdapter = new BibtexEntriesDetailViewPagerFragment.EntrySlidePagerAdapter(BibtexEntriesDetailViewPagerFragment.this.getActivity(), new ArrayList<>());
@@ -61,7 +70,7 @@ public class BibtexEntriesDetailViewPagerFragment extends Fragment {
     }
 
     private void setViewPager() {
-        List<Entry> entries = projectViewModel.getCurrentEntriesInList();
+        List<Entry> entries = getCurrentEntries();
         if(entries.size() == 0) {
             noEntriesDetails.setVisibility(View.VISIBLE);
         } else {
@@ -69,7 +78,12 @@ public class BibtexEntriesDetailViewPagerFragment extends Fragment {
             pagerAdapter = new BibtexEntriesDetailViewPagerFragment.EntrySlidePagerAdapter(BibtexEntriesDetailViewPagerFragment.this.getActivity(), entries);
             viewPager.setAdapter(pagerAdapter);
         }
-        int currentItemPosition = projectViewModel.getCurrentEntryInListCount();
+        int currentItemPosition = 0;
+        if(requireActivity() instanceof ProjectActivity) {
+            currentItemPosition = projectViewModel.getCurrentEntryInListCount();
+        } else if(requireActivity() instanceof TaxonomiesActivity) {
+            currentItemPosition = taxonomiesViewModel.getCurrentEntryInListCount();
+        }
         if(currentItemPosition != 0 && currentItemPosition < entries.size()) {
             viewPager.setCurrentItem(currentItemPosition, false);
         }
@@ -77,16 +91,34 @@ public class BibtexEntriesDetailViewPagerFragment extends Fragment {
 
     private void classifyEntry() {
         Intent intent = new Intent(getActivity(), ClassificationActivity.class);
-        Entry entry = projectViewModel.getCurrentEntriesInList().get(viewPager.getCurrentItem());
-        intent.putExtra("repo", repoId);
-        intent.putExtra("entry", entry.getEntryId());
-        startActivity(intent);
+        List<Entry> entries = getCurrentEntries();
+        int index = viewPager.getCurrentItem();
+        if(entries.size() > index) {
+            Entry entry = entries.get(index);
+            intent.putExtra("repo", repoId);
+            intent.putExtra("entry", entry.getEntryId());
+            startActivity(intent);
+        }
     }
 
     private void deleteEntry() {
-        Entry entry = projectViewModel.getCurrentEntriesInList().get(viewPager.getCurrentItem());
-        projectViewModel.deleteById(entry.getEntryId(), repoId);
-        requireActivity().onBackPressed();
+        int index = viewPager.getCurrentItem();
+        List<Entry> entries = getCurrentEntries();
+        if(entries.size() > index) {
+            Entry entry = entries.get(index);
+            projectViewModel.deleteById(entry.getEntryId(), repoId);
+            requireActivity().onBackPressed();
+        }
+    }
+
+    public List<Entry> getCurrentEntries() {
+        List<Entry> entries = new ArrayList<>();
+        if(requireActivity() instanceof ProjectActivity) {
+            entries = projectViewModel.getCurrentEntriesInList();
+        } else if(requireActivity() instanceof TaxonomiesActivity) {
+            entries = taxonomiesViewModel.getCurrentEntriesInList();
+        }
+        return entries;
     }
 
     @Override
@@ -115,17 +147,23 @@ public class BibtexEntriesDetailViewPagerFragment extends Fragment {
 
     public class EntrySlidePagerAdapter extends FragmentStateAdapter {
         List<Entry> entries;
+        FragmentActivity fa;
 
         public EntrySlidePagerAdapter(FragmentActivity fa, List<Entry> entries) {
             super(fa);
             this.entries = entries;
+            this.fa = fa;
         }
 
         @NonNull
         @Override
         public Fragment createFragment(int position) {
             Entry entry = entries.get(position);
-            projectViewModel.setCurrentEntryIdForCard(entry.getEntryId());
+            if(fa instanceof ProjectActivity) {
+                projectViewModel.setCurrentEntryIdForCard(entry.getEntryId());
+            } else if(fa instanceof TaxonomiesActivity) {
+                taxonomiesViewModel.setCurrentEntryIdForCard(entry.getEntryId());
+            }
             return new BibtexEntryDetailFragment();
         }
